requirements.txt
asgiref==3.2.3
awscli==1.16.241
botocore==1.12.231
Click==7.0
colorama==0.3.9
Django==3.0.3
django-admin==2.0.0
django-adminlte3==0.1.2
django-excel-response2==2.0.8
django-mysql-pymysql==0.1
django-six==1.0.4
docutils==0.15.2
gcloud==0.17.0
googleapis-common-protos==1.5.5
httplib2==0.12.0
jmespath==0.9.4
jws==0.1.3
oauth2client==3.0.0
protobuf==3.5.2.post1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycryptodome==3.4.3
PyMySQL==0.8.1
Pyrebase==3.0.27
python-dateutil==2.8.0
python-dotenv==0.6.3
python-jwt==2.0.1
pytz==2018.4
PyYAML==5.1.2
requests==2.11.1
requests-toolbelt==0.7.0
rsa==3.4.2
s3transfer==0.2.1
screen==1.0.1
six==1.11.0
sqlparse==0.3.0
urllib3==1.25.4
xlwt==1.3.0
mysqlclient==1.4.2.post1
env-file==2019.4.13
django-filter==2.2.0
